Created a RunTimeAIUnitGenerator
To test the RunTimeAIUnitGenerator, created a new Unit3 Prefab with other color, assigned it to the player number 1 which should be the AI. When creating the AI stuff, I can create a more difficult player AI with number 2. When assigning the Run Time AI created unit then i can assig the AI to the number 2 AI script so AI 2 can mange this unit. Remember player is number 0, AI is player 1 or more if you want to. Each of the numbers are in the Players Game object with respective scripts.