Modified the Create objectives functionality
In this commit, modifed the create objectives functionality. Now, objectives are parented to a cell in the CellGrid. Modified the SampleUnit Script to check for UnitMoved processes. If a unit moves into a cell with a objective then check if the unit was friend or foe, show debug message either way. When a unit moves to a cell with a objective then it will "Get the objective" for now it just destroys it after a delay of 2.1 seconds.