First experiment for modifing spawning in units.
In this commit, created a coroutine called EnableFriendUnitsMeshRenderers that will  enable the mesh renderers of "Friend" units after a delay of 0.5 seconds. The reason for this is becuase in the foreach loop of AddUnits function call I disabled the mesh renderer. Just for experiments sake. This will serve as baseline code ideas to implement task 1 of the In progress tab of Trello (create new CustomUnitGenerator script).
The EnableFriendUnitMeshRenderers also as a matter of experiment moves units by a new Vector3 called spawnInOffset.